Extract shared media conversion core
